# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handler.LoginRequest:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  handler.RegisterRequest:
    properties:
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - password
    - phone
    - username
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      phone:
        description: 手机号，唯一索引
        type: string
      updatedAt:
        type: string
      username:
        description: 用户名
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a sample server for Test Demo.
  title: Test Demo API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - users
  /users:
    get:
      description: 获取所有用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      summary: 获取所有用户
      tags:
      - users
  /users/{phone}:
    get:
      description: 根据手机号获取用户信息
      parameters:
      - description: 手机号
        in: path
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: 获取用户信息
      tags:
      - users
swagger: "2.0"